GH-318: added background, enabled password check
